+++ sdk/servicebus/azure-servicebus/azure/servicebus/_common/message.py
@@ -85,6 +85,8 @@ class Message(object):  # pylint: disable=too-many-public-methods,too-many-insta
     :keyword str reply_to_session_id: The session identifier augmenting the `reply_to` address.
     :keyword str encoding: The encoding for string data. Default is UTF-8.
 
+    :ivar AMQPMessage amqp_message: Advanced use only.  The internal AMQP message payload that is sent or received.
+
     .. admonition:: Example:
 
         .. literalinclude:: ../samples/sync_samples/sample_code_servicebus.py
@@ -105,6 +107,7 @@ class Message(object):  # pylint: disable=too-many-public-methods,too-many-insta
         self._amqp_header = uamqp.message.MessageHeader()
 
         if 'message' in kwargs:
+            # Note: This cannot be renamed until UAMQP no longer relies on this specific name.
             self.message = kwargs['message']
             self._amqp_properties = self.message.properties
             self._amqp_header = self.message.header
@@ -123,6 +126,8 @@ class Message(object):  # pylint: disable=too-many-public-methods,too-many-insta
             self.time_to_live = kwargs.pop("time_to_live", None)
             self.partition_key = kwargs.pop("partition_key", None)
             self.via_partition_key = kwargs.pop("via_partition_key", None)
+        # If message is the full message, amqp_message is the "public facing interface" for what we expose.
+        self.amqp_message = AMQPMessage(self.message)
 
     def __str__(self):
         return str(self.message)
@@ -1074,3 +1079,77 @@ class ReceivedMessage(ReceivedMessageBase):
         self._expiry = utc_from_timestamp(expiry[MGMT_RESPONSE_MESSAGE_EXPIRATION][0]/1000.0) # type: datetime.datetime
 
         return self._expiry
+
+
+class AMQPMessage(object):
+    """
+    The internal AMQP message that this ServiceBusMessage represents.
+
+    :param properties: Properties to add to the message.
+    :type properties: ~uamqp.message.MessageProperties
+    :param application_properties: Service specific application properties.
+    :type application_properties: dict
+    :param annotations: Service specific message annotations. Keys in the dictionary
+     must be `uamqp.types.AMQPSymbol` or `uamqp.types.AMQPuLong`.
+    :type annotations: dict
+    :param delivery_annotations: Delivery-specific non-standard properties at the head of the message.
+     Delivery annotations convey information from the sending peer to the receiving peer.
+     Keys in the dictionary must be `uamqp.types.AMQPSymbol` or `uamqp.types.AMQPuLong`.
+    :type delivery_annotations: dict
+    :param header: The message header.
+    :type header: ~uamqp.message.MessageHeader
+    :param footer: The message footer.
+    :type footer: dict
+
+    """
+    def __init__(self, message):
+        # type: (uamqp.Message) -> None
+        self._message = message
+
+    @property
+    def properties(self):
+        return self._message.properties
+
+    @properties.setter
+    def properties(self, value):
+        self._message.properties = value
+
+    @property
+    def application_properties(self):
+        return self._message.application_properties
+
+    @application_properties.setter
+    def application_properties(self, value):
+        self._message.application_properties = value
+
+    @property
+    def annotations(self):
+        return self._message.annotations
+
+    @annotations.setter
+    def annotations(self, value):
+        self._message.annotations = value
+
+    @property
+    def delivery_annotations(self):
+        return self._message.delivery_annotations
+
+    @delivery_annotations.setter
+    def delivery_annotations(self, value):
+        self._message.delivery_annotations = value
+
+    @property
+    def header(self):
+        return self._message.header
+
+    @header.setter
+    def header(self, value):
+        self._message.header = value
+
+    @property
+    def footer(self):
+        return self._message.footer
+
+    @footer.setter
+    def footer(self, value):
+        self._message.footer = value